+++ NEWS.md
@@ -3,6 +3,7 @@
 
 ## enhancements
 - `giottoLargeImage` `max_window` and `colors` slot info is now followed during ggplot plotting
+- `giottoAffineImage` compatibility for giotto ggplot2 plotting functions
 
 ## new
 - `geom_text_repel()` and `geom_label_repel()` from `ggplot2` are now re-exported